# orgnote-backend/app/docs/swagger.yaml
definitions:
  FileChange:
    description: File change information for synchronization
    properties:
      contentHash:
        example: a1b2c3d4e5f6...
        type: string
      deleted:
        example: false
        type: boolean
      deletedAt:
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      path:
        example: notes/todo.org
        type: string
      size:
        example: 1024
        type: integer
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      version:
        example: 1
        type: integer
    required:
    - deleted
    - id
    - path
    - updatedAt
    - version
    type: object
  FileUploadResponse:
    description: Response after successful file upload
    properties:
      contentHash:
        example: a1b2c3d4e5f6...
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      path:
        example: notes/todo.org
        type: string
      size:
        example: 1024
        type: integer
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      uploaded:
        example: true
        type: boolean
      version:
        example: 1
        type: integer
    required:
    - contentHash
    - id
    - path
    - size
    - updatedAt
    - uploaded
    - version
    type: object
  SyncChangesResponse:
    description: Response containing file changes since last sync
    properties:
      changes:
        items:
          $ref: '#/definitions/FileChange'
        type: array
      cursor:
        example: 507f1f77bcf86cd799439011
        type: string
      hasMore:
        example: false
        type: boolean
      serverTime:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - changes
    - hasMore
    - serverTime
    type: object
  VersionConflictResponse:
    description: Response when optimistic locking fails due to version mismatch
    properties:
      error:
        example: version mismatch
        type: string
      path:
        example: notes/todo.org
        type: string
      serverVersion:
        example: 5
        type: integer
    required:
    - error
    - path
    - serverVersion
    type: object
  handlers.HttpError-any:
    properties:
      data: {}
      message:
        type: string
    type: object
  handlers.HttpResponse-FileUploadResponse-any:
    properties:
      data:
        $ref: '#/definitions/FileUploadResponse'
      meta: {}
    type: object
  handlers.HttpResponse-SyncChangesResponse-any:
    properties:
      data:
        $ref: '#/definitions/SyncChangesResponse'
      meta: {}
    type: object
...
  handlers.HttpResponse-models_FileMetadata-any:
    properties:
      data:
        $ref: '#/definitions/models.FileMetadata'
      meta: {}
    type: object
...
  handlers.SystemInfo:
    properties:
      environment:
        $ref: '#/definitions/models.EnvironmentInfo'
      update:
        $ref: '#/definitions/models.OrgNoteClientUpdateInfo'
    type: object
...
  models.EnvironmentInfo:
    properties:
      minClientVersion:
        type: string
      selfHosted:
        type: boolean
    type: object
  models.FileMetadata:
    properties:
      contentHash:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      path:
        type: string
      size:
        type: integer
      updatedAt:
        type: string
      userId:
        type: string
      version:
        type: integer
    type: object
  models.OrgNoteClientUpdateInfo:
    properties:
      changeLog:
        type: string
      url:
        type: string
      version:
        type: string
    type: object
...
info:
  contact:
    email: artawower@protonmail.com
    name: API Support
  description: List of methods for work with Org Note.
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  termsOfService: http://swagger.io/terms/
  title: Org Note API
  version: 0.0.1
paths:
...
  /sync/changes:
    get:
      consumes:
      - application/json
      description: Returns file changes since the specified timestamp
      parameters:
      - description: Unix timestamp in milliseconds for incremental sync
        in: query
        name: since
        type: integer
      - description: 'Maximum number of changes to return (default: 100, max: 500)'
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HttpResponse-SyncChangesResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Get file changes
      tags:
      - sync
  /sync/files:
    delete:
      consumes:
      - application/json
      description: Soft delete a file (creates tombstone for sync)
      parameters:
      - description: File path
        in: query
        name: path
        required: true
        type: string
      - description: Expected version for optimistic locking
        in: query
        name: version
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HttpResponse-models_FileMetadata-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/VersionConflictResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Delete a file
      tags:
      - sync
    get:
      description: Download file content by path
      parameters:
      - description: File path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          headers:
            X-Content-Hash:
              description: SHA-256 hash of file content
              type: string
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Download a file
      tags:
      - sync
    put:
      consumes:
      - multipart/form-data
      description: Upload a file to sync storage with content-addressable deduplication
      parameters:
      - description: Relative file path
        in: formData
        name: filePath
        required: true
        type: string
      - description: File content
        in: formData
        name: file
        required: true
        type: file
      - description: Expected version for optimistic locking
        in: formData
        name: expectedVersion
        type: integer
      - description: SHA-256 hash for verification
        in: header
        name: X-Content-Hash
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HttpResponse-FileUploadResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/VersionConflictResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Upload a file
      tags:
      - sync
  /system-info/{version}:
    get:
      consumes:
      - application/json
      description: Get system info
      parameters:
      - description: provider
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SystemInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: GetSystemInfo
      tags:
      - system info
  /system-info/client-update/{version}:
    get:
      consumes:
      - application/json
      parameters:
      - description: provider
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrgNoteClientUpdateInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: GetUpdatesFromVersion
      tags:
      - system info
  /system-info/client-update/latest:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrgNoteClientUpdateInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: GetLatestClientUpdate
      tags:
      - system info
...
swagger: "2.0"